Разбивка по модулям. Сделано: get_companies_and_vacancies_count, get_all_vacancies у DBManager
